requirements.txt
google-api-python-client==2.118.0
pandas==2.2.0
pyarrow==15.0.0
python-dotenv==1.0.1
pytest==8.0.0
google-cloud-storage==2.14.0
requests==2.31.0